Add files via upload
Ya todo compila. 
Para CLONAR el grafo, lo que hice fué usar el método clone() que está en Digrafo. Ahora, no se que hacía exactamente simetria(), creo que era lo mismo; de todas maneras lo dejé allí.
Ahora en los Arcos y Vertice los pesos van a ser int, porque no va a haber alturas double, por lo que se...

+++ Grafo.java
@@ -17,7 +17,7 @@ public interface Grafo
     
     public boolean agregarVertice(Vertice v);
 
-    public boolean agregarVertice(String id, double peso);
+    public boolean agregarVertice(String id, int peso);
     
     public Vertice obtenerVertice(String id);
